Move entrypoint to diff file to make e2e testing easy
